+++ tests/test_cmd2.py
@@ -1620,7 +1620,8 @@ def test_get_settable_completion_items(base_app):
 
 def test_alias_no_subcommand(base_app):
     out, err = run_cmd(base_app, 'alias')
-    assert "Usage: alias [-h]" in out[0]
+    assert "Usage: alias [-h]" in err[0]
+    assert "Error: the following arguments are required: subcommand" in err[1]
 
 def test_alias_create(base_app):
     # Create the alias
@@ -1713,7 +1714,8 @@ def test_multiple_aliases(base_app):
 
 def test_macro_no_subcommand(base_app):
     out, err = run_cmd(base_app, 'macro')
-    assert "Usage: macro [-h]" in out[0]
+    assert "Usage: macro [-h]" in err[0]
+    assert "Error: the following arguments are required: subcommand" in err[1]
 
 def test_macro_create(base_app):
     # Create the macro